[# 172] 은행 정보 조회 기능 구현

+++ airbng-pay/src/main/java/com/airbng/pay/dto/AccountRegisterRequest.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class AccountRegisterRequest {
-    private Long bankCode;
+    private Integer bankCode;
     private String accountNumber;
     private String holderName;
 }
